moved config object to separate module
