+++ config/config.py
@@ -1,0 +1,3 @@
+from config.sah_config import SAHConfig, get_db_credentials_path
+
+sah_config = SAHConfig(credentials_path=get_db_credentials_path())
