+++ controllers/filters.py
@@ -4,7 +4,7 @@ from quart import Blueprint, Response, abort, jsonify, request
 from sqlalchemy import select
 
 from auth import UserData, requires_auth
-from config import sah_config
+from config.config import sah_config
 
 from .common import validator
 from models import Filter
