+++ create_app.py
@@ -31,7 +31,7 @@ from quart import Quart, Response, jsonify
 from quart_cors import cors
 
 from auth import AuthError
-from config import sah_config
+from config.config import sah_config
 from controllers import routers
 
 from utils.validator import ValidationError
